django/src/requirements.txt
Django==4.2
psycopg2
mysqlclient
djangorestframework
djangorestframework-simplejwt